- copy_wwise_events no longer automatically saves - rewrote extremly inefficient cleanup task in import_wems - Removed save/repack buttons from mass transfer

+++ yonder/gui/widgets/yay.py
@@ -47,5 +47,5 @@ def yay(fadeout: float = 2.0):
 
     # Make sure this gets deleted even if our callback is somehow lost
     dpg.set_frame_callback(
-        dpg.get_frame_count() + fadeout * 60, fadeout_cb
+        dpg.get_frame_count() + int(fadeout * 60), fadeout_cb
     )
